pyproject.toml
[project]
name = "nexus-api"
version = "0.1.0"
description = ""
authors = [
    {name = "MarcosCast",email = "marcosmax.5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
